+++ NightVisionShaders [Java]/current/Night Vision Shaders [Java]/changelog.md
@@ -1,1 +1,1 @@
-- Updated supported sodium versions
+- Updated to sodium 0.7.0 and minecraft 1.21.9